% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_genotyping_accuracy.R
\name{calc_genotyping_accuracy}
\alias{calc_genotyping_accuracy}
\title{Calculate genotyping accuracy from RNA-seq calls.}
\usage{
calc_genotyping_accuracy(data_x, data_y, by_locus = TRUE)
}
\arguments{
\item{data_x}{data.frame.}

\item{data_y}{data.frame. Gold standard data.}

\item{by_locus}{logical. Whether or not to summarize by locus.}
}
\value{
A data.frame.
}
\description{
Calculates genotyping accuracies from RNA-seq calls using a gold standard.
}
