% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts_to_fpkm.R
\name{counts_to_fpkm}
\alias{counts_to_fpkm}
\title{Convert read counts to FPKM}
\usage{
counts_to_fpkm(counts, eff_len)
}
\arguments{
\item{counts}{numeric.}

\item{eff_len}{numeric.}
}
\value{
A numeric vector of FPKM values.
}
\description{
Convert read counts to FPKM
}
