% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts_to_tpm.R
\name{counts_to_tpm}
\alias{counts_to_tpm}
\title{Convert read counts to TPM}
\usage{
counts_to_tpm(counts, eff_len)
}
\arguments{
\item{counts}{numeric.}

\item{eff_len}{numeric.}
}
\value{
A numeric vector of TPM values.
}
\description{
Convert read counts to TPM
}
