% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpkm_to_tpm.R
\name{fpkm_to_tpm}
\alias{fpkm_to_tpm}
\title{Convert FPKM to TPM}
\usage{
fpkm_to_tpm(fpkm)
}
\arguments{
\item{fpkm}{numeric.}
}
\value{
A numeric vector of TPM values.
}
\description{
Convert FPKM to TPM
}
