% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hla_apply_zigosity_threshold.R
\name{hla_apply_zigosity_threshold}
\alias{hla_apply_zigosity_threshold}
\title{Apply zigosity threshold to expression estimates}
\usage{
hla_apply_zigosity_threshold(hla_quants, th = 0.1)
}
\arguments{
\item{hla_quants}{data.frame.}

\item{th}{numeric.}
}
\value{
A data.frame
}
\description{
Apply zigosity threshold to expression estimates
}
