% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hla_compile_index.R
\name{hla_compile_index}
\alias{hla_compile_index}
\title{Process HLA alignment data in *.nuc files and return a data.frame.}
\usage{
hla_compile_index(locus, imgt.database)
}
\arguments{
\item{locus}{character string.}

\item{imgt.database}{character string.}
}
\value{
A data.frame.
}
\description{
Process HLA alignment data in *.nuc files and return a data.frame.
}
