% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hla_consensus.R
\name{hla_consensus}
\alias{hla_consensus}
\title{Create consensus sequences for a list of HLA alleles.}
\usage{
hla_consensus(hla_mat, fields, cores = 1)
}
\arguments{
\item{hla_mat}{matrix.}

\item{fields}{integer. Resolution of alleles in fields.}

\item{cores}{integer. Number of cores if running in parallel.}
}
\value{
A matrix of consensus sequences. Ambiguous positions contain the
occurring bases separated by "/".
}
\description{
\code{hla_consensus()} takes a matrix of HLA sequences and returns a matrix 
of consensus sequences at the resolution specified by the argument
\code{field}.
}
