% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hla_genotype.R
\name{hla_genotype}
\alias{hla_genotype}
\title{Infer HLA genotypes from RNA-seq quantifications.}
\usage{
hla_genotype(hla_quants, th = 0.05)
}
\arguments{
\item{hla_quants}{data.frame.}

\item{th}{numeric. Threshold for expression 2nd/1st allele.}
}
\value{
A data.frame.
}
\description{
Infer HLA genotypes from RNA-seq quantifications.
}
