% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hla_genotype_dt.R
\name{hla_genotype_dt}
\alias{hla_genotype_dt}
\title{Infer HLA genotypes from RNA-seq quantifications.}
\usage{
hla_genotype_dt(hla_quants, th = 0.05)
}
\arguments{
\item{hla_quants}{data.frame. Kallisto output data (single or multiple inds).}

\item{th}{numeric. Threshold for expression 2nd/1st allele.}
}
\value{
A data.frame
}
\description{
Infer HLA genotypes from RNA-seq quantifications.
}
