% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hla_matrix_to_list.R
\name{hla_matrix_to_list}
\alias{hla_matrix_to_list}
\title{Convert a matrix of sequences in a list}
\usage{
hla_matrix_to_list(hla_mat)
}
\arguments{
\item{hla_mat}{matrix. A matrix such the one returned by \code{hla_parsenuc}.}
}
\value{
a list.
}
\description{
Convert a matrix of sequences, such that returned by hla_parsenuc, into a
list in which each element is the collapse sequence of each allele.
}

