% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hla_minres.R
\name{hla_minres}
\alias{hla_minres}
\title{Reduce a vector of ambiguous alleles to the maximum resolution that includes
all of them.}
\usage{
hla_minres(allele_vec)
}
\arguments{
\item{allele_vec}{character vector.}
}
\value{
character vector. A single allele at the resolution that includes all
alleles in the original vector, \code{NULL} when there is no resolution that
includes all alleles.
}
\description{
Reduce a vector of ambiguous alleles to the maximum resolution that includes
all of them.
}
