% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hla_parsenuc.R
\name{hla_parsenuc}
\alias{hla_parsenuc}
\title{Process HLA alignment data in *.nuc files.}
\usage{
hla_parsenuc(locus, align_dir = getwd(), exons = NULL, omit_suffix = NULL,
  undash = TRUE, rm_dots_method = c("all", "ref"))
}
\arguments{
\item{locus}{character string.}

\item{align_dir}{character string. Path to nuc file as a character string. 
Default is working directory.}

\item{exons}{integer. Range of vectors to select (default = all exons).}

\item{omit_suffix}{character string. String of optional suffixes indicating
expression status. Possible values are NULL (default, keep all allele names
with suffixes) or a string of suffixes to omit (e.g., "NQL").}

\item{undash}{logical. Should dashes be replaced with the base present in the
reference allele?}

\item{rm_dots_method}{character string. Method used to remove dots in the
sequence ("all" to remove only dots present in all alleles or "ref" to
remove those present in the first allele on the nuc file).}
}
\value{
A matrix with the sequences from all alleles.
}
\description{
Reads in nuc files with aligned sequences and processes them, returning a
matrix in which each row is the complete sequence for an allele.
}

