% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hla_read_alignment.R
\name{hla_read_alignment}
\alias{hla_read_alignment}
\title{Process HLA alignment data in *.nuc files and return a data.frame.}
\usage{
hla_read_alignment(locus, imgtdb, exons = NULL)
}
\arguments{
\item{locus}{character string. The HLA locus to be processed (e.g., "A", "DRB1").}

\item{imgtdb}{character string. Path to the IMGTHLA directory.}

\item{exons}{integer. Range of exons (default is all exons).}
}
\value{
A data.frame.
}
\description{
Process HLA alignment data in *.nuc files and return a data.frame.
}
\examples{
\dontrun{
hla_read_alignment(locus = "DRB1", imgtdb = "~/IMGTHLA") 
}
\dontrun{
hla_read_alignment(locus = "A", imgtdb = "~/IMGTHLA", exons = c(2, 3))
}

}
