% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hla_read_utr.R
\name{hla_read_utr}
\alias{hla_read_utr}
\title{Process HLA alignment data in *.nuc files and return a data.frame.}
\usage{
hla_read_utr(gen_file, omit_suffix = "N")
}
\arguments{
\item{gen_file}{character string. Path to IMGT alignment file.}

\item{omit_suffix}{character string. String of optional suffixes indicating
expression status. Possible values are NULL (keep all allele names with
suffixes) or a string of suffixes to omit (e.g., "NQL"). Default is "N".}
}
\value{
A data.frame.
}
\description{
Process HLA alignment data in *.nuc files and return a data.frame.
}
