% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hla_readtx.R
\name{hla_readtx}
\alias{hla_readtx}
\title{Read hlaTX output file}
\usage{
hla_readtx(x)
}
\arguments{
\item{x}{Character string. Path to hlatx output file.}
}
\value{
A data.frame.
}
\description{
Read hlaTX output file
}
