% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hla_refdist.R
\name{hla_refdist}
\alias{hla_refdist}
\title{Calculate edit distances to the reference HLA allele.}
\usage{
hla_refdist(hla_mat, ref_allele, stat = c("count", "proportion"))
}
\arguments{
\item{hla_mat}{matrix. A matrix returned by \code{hla_parsenuc}.}

\item{ref_allele}{character string. The name of the reference allele in the 
same format of rownames(hla_mat).}

\item{stat}{character string. Return absulute number of differences ("count")
or proportion ("proportion")}
}
\value{
A vector of edit distances of each allele in respect to the reference 
allele.
}
\description{
Calculate edit distances to the reference HLA allele.
}
