% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hla_trimnames.R
\name{hla_trimnames}
\alias{hla_trimnames}
\title{Trim HLA allele names to the specified number of fields.}
\usage{
hla_trimnames(alleles_vec, fields = 3)
}
\arguments{
\item{alleles_vec}{character string. Vector of allele names.}

\item{fields}{integer. Number of fields to trim names to.}
}
\value{
All or unique HLA alleles trimmed names.
}
\description{
Given a vector of HLA allele names with the format "A*01:01:01:01", returns
all or unique allele names at the digit resolution specified by 
\code{fields}.
}
\examples{
hla_trimnames(c("A*01:01:01:01", "A*02:01:01:01"), fields = 2)
}
