% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase_hla.R
\name{phase_hla}
\alias{phase_hla}
\title{Phase HLA alleles given distances to 1000G haplotypes}
\usage{
phase_hla(x)
}
\arguments{
\item{x}{data.frame. A data.frame with distances of alleles to each 1000G
haplotype.}
}
\value{
A data.frame with 2 rows with inferred haplotypes for each
individual.
}
\description{
Phase HLA alleles given distances to 1000G haplotypes
}
