% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_phase_probs.R
\name{read_phase_probs}
\alias{read_phase_probs}
\title{Read in PHASE output file}
\usage{
read_phase_probs(phase_out)
}
\arguments{
\item{phase_out}{character string. Path to PHASE output file.}
}
\value{
A data.frame.
}
\description{
Reads PHASE output file and returns a data.frame with the phasing
probabilities.
}
