% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qtltools.R
\name{read_qtltools}
\alias{read_qtltools}
\title{Read in QTLtools output from secondary chunk analysis}
\usage{
read_qtltools(qtltools_output)
}
\arguments{
\item{qtltools_output}{character string. Path to file.}
}
\value{
data.frame
}
\description{
Read in QTLtools output from secondary chunk analysis
}
