% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qtltools_rtc.R
\name{read_qtltools_rtc}
\alias{read_qtltools_rtc}
\title{Read in QTLtools rtc output}
\usage{
read_qtltools_rtc(rtc_output)
}
\arguments{
\item{rtc_output}{character string. Path to file.}
}
\value{
data.frame
}
\description{
Read in QTLtools rtc output
}
