% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_phase_input.R
\name{write_phase_input}
\alias{write_phase_input}
\title{Write input for PHASE}
\usage{
write_phase_input(genotypes, outfile)
}
\arguments{
\item{genotypes}{data.frame. A data.frame with cols subject, locus, allele.}

\item{outfile}{character string. Path to output file.}
}
\value{
No return.
}
\description{
Takes a data.frame of HLA genotypes and writes a input file for PHASE
}
