% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hla_infer.R
\name{hla_infer}
\alias{hla_infer}
\title{Infer sequences missing on the IMGT alignments.}
\usage{
hla_infer(hla_df, cores = 1)
}
\arguments{
\item{hla_df}{data.frame. A data.frame such as the one returned by 
\code{hla_dataframe}.}

\item{cores}{integer. Number of cores.}
}
\value{
A data.frame with the original sequences of the complete alleles
and the inferred sequences for incomplete alleles.
}
\description{
Infer sequences missing on the IMGT alignments.
}
