% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hla_read_alignment.R
\name{hla_read_alignment}
\alias{hla_read_alignment}
\title{Process HLA alignment data in *.nuc files and return a data.frame.}
\usage{
hla_read_alignment(nuc_file, exons = NULL, by_exon = FALSE,
  omit_suffix = NULL, rm_incomplete = FALSE, keep_sep = FALSE)
}
\arguments{
\item{nuc_file}{character string. Path to IMGT alignment file.}

\item{exons}{integer. Range of vectors to select (default = all exons).}

\item{by_exon}{logical. Whether to return data separated by exons.}

\item{omit_suffix}{character string. String of optional suffixes indicating
expression status. Possible values are NULL (default, keep all allele names
with suffixes) or a string of suffixes to omit (e.g., "NQL").}

\item{rm_incomplete}{logical. Whether to remove incomplete sequences.}

\item{keep_sep}{logical. Whether to keep the exon separator character "|" in
the sequences.}
}
\value{
A data.frame.
}
\description{
Process HLA alignment data in *.nuc files and return a data.frame.
}

