% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_phase.R
\name{read_phase}
\alias{read_phase}
\title{Read in PHASE output file}
\usage{
read_phase(phase_out, loci)
}
\arguments{
\item{phase_out}{character string. Path to PHASE output file.}

\item{loci}{character string. Loci names in genomic order.}
}
\value{
A data.frame.
}
\description{
Reads PHASE output file and returns a data.frame with best haplotypes per
individual
}
