% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qtltools_rtc.R
\name{read_qtltools_rtc}
\alias{read_qtltools_rtc}
\title{Read in QTLtools rtc output}
\usage{
read_qtltools_rtc(rtc_output, col_names = TRUE)
}
\arguments{
\item{rtc_output}{character string. Path to file.}

\item{col_names.}{Whether the rtc output has col names. Normal run does, when
a genomic region is selected, it does not.}
}
\value{
data.frame
}
\description{
Read in QTLtools rtc output
}
